+++ datanode/sqlstore/stop_orders_test.go
@@ -60,6 +60,7 @@ func generateTestStopOrders(t *testing.T, testOrders []testStopOrderInputs) []en
 				Type:        entities.OrderTypeMarket,
 				Reference:   o.orderID,
 			},
+			RejectionReason: entities.StopOrderRejectionReasonUnspecified,
 		}
 		orders = append(orders, so)
 	}