+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - [8120](https://github.com/vegaprotocol/vega/issues/8120) - Data node `API` support for spot markets, data and governance.
 - [8762](https://github.com/vegaprotocol/vega/issues/8762) - Data node `API` support for perpetual markets, data and governance.
 - [8761](https://github.com/vegaprotocol/vega/issues/8761) - Add terminating `PERPS` via governance.
+- [9021](https://github.com/vegaprotocol/vega/issues/9021) - Add rejection reason to stop orders.
 
 ### 🐛 Fixes